ft-vlm/pyproject.toml
[project]
name = "ft-vlm"
version = "0.1.0"
description = ""
authors = [
    {name = "kuro2907",email = "nam.le@visily.ai"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.44.0",
    "trl>=0.9.6",
    "datasets>=2.19.0",
    "bitsandbytes>=0.43.0; platform_system != 'Darwin'",
    "peft>=0.11.1",
    "qwen-vl-utils>=0.0.11",
    "accelerate>=0.33.0",
    "wandb>=0.17.0",
    "pillow>=10.0.0",
]

[tool.poetry]
packages = [{include = "ft_vlm", from = "src"}]

[project.scripts]
ft-vlm-train = "ft_vlm.fine_tuning.train:cli_main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
